Move indexing tests into unified tests/test_indexing.py
